Add files via upload
This is updated code by Adi Natan 
The modifications:
1. The Y matrix code is now vectorized making it ~ x20 faster. 
2. The code has an optional zero padding input similar to Matlab's fft functionality.
3. The code supports also arrays not only vectors similar to Matlab's fft functionality, meaning that the Bessel zeros and Y matrix are only calculated once per run.
4.  The code now use an updated Bessel zeros calculation code (by Jason Nicholson).
 
Adi also noted that in the original “Guide to DHT” and the YmatrixAssembly function code there was a variable called ‘zeros’, which is incidentally also a name of a Matlab built-in function zeros. This can be a source of bugs if used in addition to other matlab code, and may confuse users.